requirements/ci.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    ./bin/compile_dependencies.sh
#
asgiref==3.5.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
beautifulsoup4==4.11.1
    # via webtest
certifi==2022.6.15
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
charset-normalizer==2.1.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
coverage==4.5.4
    # via -r requirements/test-tools.in
django==3.2.15
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-axes
    #   django-rosetta
django-axes==5.39.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-cleanup==6.0.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-ipware==4.0.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-axes
django-rosetta==0.9.8
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
django-webtest==1.9.10
    # via -r requirements/test-tools.in
factory-boy==3.2.1
    # via -r requirements/test-tools.in
faker==14.1.0
    # via factory-boy
freezegun==1.2.2
    # via -r requirements/test-tools.in
idna==3.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
pillow==9.2.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
polib==1.1.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-rosetta
psycopg2==2.9.3
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
python-decouple==3.6
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
python-dotenv==0.20.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
python-magic==0.4.27
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
pytz==2022.2.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
requests==2.28.1
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-rosetta
    #   responses
responses==0.21.0
    # via -r requirements/test-tools.in
six==1.16.0
    # via python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django
urllib3==1.26.11
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   requests
    #   responses
uwsgi==2.0.20
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
waitress==2.1.2
    # via webtest
webob==1.8.7
    # via webtest
webtest==3.0.0
    # via django-webtest

# The following packages are considered to be unsafe in a requirements file:
setuptools==65.1.0
    # via
    #   -c requirements/base.txt
    #   -r requirements/base.txt
    #   django-axes
